Merged PR 469: [Other] use BodyRuleType enums instead of string conversions
## Purpose
<!-- why are we doing this? -->

*Task description start*
body-rule.type.ts: BodyEquality and BodyContains enums need to change to numbers which correspond to the same numbers on the server enums. Make the BodyRuleType enum the same as the one on the server.
body-rule-list-item.component.ts: the getBodyType() method needs to return a string representing the "name" of the enum (so for example if BodyContains is passed in as the bodyRule.type then it should return "Body Contains") as this is used for the dropdown
orbital-admin.service.ts: the exportMockDefinition method's code to change the request match rules body rules from a string into a number can be removed. When this is removed, it will be clear that the outer foreach loop isn't doing anything, so that can be removed too.
*Task description end*

The BodyRuleType enums are re-re-converted into strings and into numbers throughout the application, causing code bloat. Only use numbers and convert when displaying to the user. This also fixes a bug when importing from the server causes the body types to show up as numbers in the list (they should be strings) and fixes the bug where duplicate rules can be added.

## Changes
<!-- high level summary in bullet points -->
<!-- annotate code where workarounds are being used -->

- remove strings for the BodyRuleType enums (replace with numbers as used on the server)
- remove re-re-conversion code in OrbitalAdminService

## Tests
<!-- will the existing tests cover these new changes? -->

- tests were added to getBodyRuleType() as it now returns a string instead of the enum directly (because it has to convert it)

## Pre-review checklist:
- [x] builds; tests pass, including integration tests
- [x] is based off the latest branch
- [x] has a meaningful title and is tagged with [Test], [Bug], [Feature], [Refactor], or [Other]
- [x] has a related issue
- [x] is indivisible
- [x] is DRY
- [x] clear naming and spell checked; commented if reviewer did not understand
- [ ] if this is related to the UI, screenshots/GIFs are provided

<!-- delete all non-related auto-attached work items in the right-hand sidebar after creating →
<!-- if items in the backlog are created as a result of this PR (e.g. a suggested refactor), link the work items here using the #<work item number> syntax so that it links to the issue -->

Related work items: #1597