Merged PR 683: Workaround for concurrent Mockdefinitions not saving: lock database while accessing
When exporting four or more mocks from the designer, the server will randomly not save some of the mocks, but the designer will report that the export was successful. This causes data loss. This appears to be caused by the database not being safe when writing to it from multiple threads.

Locking the database on every request is *not* an optimal solution as this limits throughput and is not considered good practise. Additionally, locking in this manner can cause deadlocks if a process gets stuck on a lock, which can cause difficult-to-debug issues. This method was chosen so that the underlying database was not touched and to ensure compatibility. However, the database will be replaced in an upcoming version with one which does not require locking.

When the database is replaced, the locks should be removed. If the locks are still required, then the database should not be used as it is not ACID compliant.

Related work items: #2086